fix: move handleJoin to SpaceBanner
